app/requirements.txt
anyio==4.8.0
azure-core==1.32.0
azure-identity==1.20.0
black==23.7.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.12
cryptography==44.0.1
ecdsa==0.19.0
fastapi==0.95.0
flake8==6.1.0
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.67.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
msal==1.31.1
msal-extensions==1.2.0
mypy==1.5.1
mypy-extensions==1.0.0
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==1.10.8
pyflakes==3.1.0
PyJWT==2.10.1
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.0.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.46
starlette==0.26.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.21.1
